decisions: CORRECTION — the water-world is the Phase-1 promise kept, not a contradiction
Joseph caught it: the ordinum forbids land as an initial condition (abyssal
charge[emergent-land] :tag gate — 'delivered by uplift/proto-tectonic processes,
NEVER an initial condition'), and phase[protogenic] promises a water-covered
surface. So the sea_level_probe CONFIRMS the ordinum and CONVICTS the spine —
SEA_LEVEL_M=4000 is manufacturing forbidden land. Bimodality is not a fix to the
prior; isostasy is the MECHANISM by which Abyssal DELIVERS land. Erosion no-op'ing
on submarine ground is physically correct (erosion-carving is itself a phase-3 gate).

Second over-interpretation of a correct measurement today — logged as a standing
guard: check the LADDER, not modern Earth.

+++ crates/vivarium-world/examples/grid_lab/mesh.rs
@@ -69,8 +69,8 @@ pub struct Edge {
     /// Outward unit normal to the shared edge, in the tangent plane at the mid-edge.
     pub normal: V3,
     /// The two endpoints of the shared edge.
-    pub va: usize,
-    pub vb: usize,
+    pub va: u32,
+    pub vb: u32,
 }
 
 /// Everything a grid must be able to say about itself.
